+++ CHANGELOG.md
@@ -3,3 +3,8 @@
 ## 0.0.1
 
 - Initial version
+
+## 0.1.0
+
+- Add build and rebuild methods to BuiltList, BuiltMap, BuiltSet.
+- Add update methods to ListBuilder, MapBuilder, SetBuilder.